+++ src/Simplify.Web.Tests/Controllers/V2/Metadata/MetadataFactoryTests/Controller2MetadataFactoryTests.cs
@@ -65,7 +65,7 @@ public class Controller2MetadataFactoryTests
 		Assert.That(metaData.ExecParameters, Is.Not.Null);
 		Assert.That(metaData.ExecParameters!.RunPriority, Is.EqualTo(1));
 
-		Assert.That(metaData.ExecParameters.Routes.Count, Is.EqualTo(6));
+		Assert.That(metaData.ExecParameters.Routes.Count, Is.EqualTo(7));
 
 		var firstControllerRoute = metaData.ExecParameters.Routes.First(x => x.Key == HttpMethod.Get);
 
@@ -83,5 +83,6 @@ public class Controller2MetadataFactoryTests
 		Assert.That(execParameters.Routes[HttpMethod.Patch].Path, Is.EqualTo("/test-action3"));
 		Assert.That(execParameters.Routes[HttpMethod.Delete].Path, Is.EqualTo("/test-action4"));
 		Assert.That(execParameters.Routes[HttpMethod.Options].Path, Is.EqualTo("/test-action5"));
+		Assert.That(execParameters.Routes[HttpMethod.Query].Path, Is.EqualTo("/test-action6"));
 	}
 }